+++ packages/expo-dev-launcher/ios/EXDevLauncherBridgeDelegate.mm
@@ -32,7 +32,11 @@ static NSString *const kRNConcurrentRoot = @"concurrentRoot";
     enableTM = YES;
 #endif
 
+#if REACT_NATIVE_MINOR_VERSION >= 74
+    RCTAppSetupPrepareApp(application, enableTM, facebook::react::EmptyReactNativeConfig());
+#else
     RCTAppSetupPrepareApp(application, enableTM);
+#endif
 
     self.bridge = [super createBridgeAndSetAdapterWithLaunchOptions:launchOptions];