+++ packages/expo-dev-menu/ios/Tests/DevMenuAppInstanceTest.swift
@@ -51,11 +51,10 @@ class DevMenuAppInstanceTest: QuickSpec {
         bridge: mockedBridge
       )
 
-        let extraModules = appInstance.extraModules(for: mockedBridge)
+      let extraModules = appInstance.extraModules(for: mockedBridge)
 
-      expect(extraModules).toNot(beNil())
-      expect(extraModules?.first { type(of: $0).moduleName() == "DevLoadingView" }).toNot(beNil())
-      expect(extraModules?.first { type(of: $0).moduleName() == "DevSettings" }).toNot(beNil())
+      expect(extraModules.first { type(of: $0).moduleName() == "DevLoadingView" }).toNot(beNil())
+      expect(extraModules.first { type(of: $0).moduleName() == "DevSettings" }).toNot(beNil())
     }
   }
 }